+++ Sources/SwiftPhoenixClient/Socket.swift
@@ -36,10 +36,10 @@ public typealias PayloadClosure = () -> Payload?
 
 /// Struct that gathers callbacks assigned to the Socket
 struct StateChangeCallbacks {
-  var open: [(ref: String, callback: Delegated<Void, Void>)] = []
-  var close: [(ref: String, callback: Delegated<Void, Void>)] = []
-  var error: [(ref: String, callback: Delegated<Error, Void>)] = []
-  var message: [(ref: String, callback: Delegated<Message, Void>)] = []
+    let open: SynchronizedArray<(ref: String, callback: Delegated<URLResponse?, Void>)> = .init()
+    let close: SynchronizedArray<(ref: String, callback: Delegated<(Int, String?), Void>)> = .init()
+    let error: SynchronizedArray<(ref: String, callback: Delegated<(Error, URLResponse?), Void>)> = .init()
+    let message: SynchronizedArray<(ref: String, callback: Delegated<Message, Void>)> = .init()
 }
 
 
@@ -78,8 +78,8 @@ public class Socket: PhoenixTransportDelegate {
     return self.paramsClosure?()
   }
   
-  /// The optional params closure used to get params whhen connecting. Must
-  /// be set when initializaing the Socket.
+  /// The optional params closure used to get params when connecting. Must
+  /// be set when initializing the Socket.
   public let paramsClosure: PayloadClosure?
   
   /// The WebSocket transport. Default behavior is to provide a
@@ -92,14 +92,20 @@ public class Socket: PhoenixTransportDelegate {
   /// Override to provide custom encoding of data before writing to the socket
   public var encode: (Any) -> Data = Defaults.encode
   
-  /// Override to provide customd decoding of data read from the socket
+  /// Override to provide custom decoding of data read from the socket
   public var decode: (Data) -> Any? = Defaults.decode
   
   /// Timeout to use when opening connections
   public var timeout: TimeInterval = Defaults.timeoutInterval
+    
+  /// Custom headers to be added to the socket connection request
+  public var headers: [String : Any] = [:]
   
   /// Interval between sending a heartbeat
   public var heartbeatInterval: TimeInterval = Defaults.heartbeatInterval
+
+  /// The maximum amount of time which the system may delay heartbeats in order to optimize power usage
+  public var heartbeatLeeway: DispatchTimeInterval = Defaults.heartbeatLeeway
   
   /// Interval between socket reconnect attempts, in seconds
   public var reconnectAfter: (Int) -> TimeInterval = Defaults.reconnectSteppedBackOff
@@ -134,14 +140,14 @@ public class Socket: PhoenixTransportDelegate {
   // MARK: - Private Attributes
   //----------------------------------------------------------------------
   /// Callbacks for socket state changes
-  var stateChangeCallbacks: StateChangeCallbacks = StateChangeCallbacks()
+  let stateChangeCallbacks: StateChangeCallbacks = StateChangeCallbacks()
   
   /// Collection on channels created for the Socket
-  var channels: [Channel] = []
+  public internal(set) var channels: [Channel] = []
   
   /// Buffers messages that need to be sent once the socket has connected. It is an array
   /// of tuples, with the ref of the message to send and the callback that will send the message.
-  var sendBuffer: [(ref: String?, callback: () throws -> ())] = []
+  let sendBuffer = SynchronizedArray<(ref: String?, callback: () throws -> ())>()
   
   /// Ref counter for messages
   var ref: UInt64 = UInt64.min // 0 (max: 18,446,744,073,709,551,615)
@@ -201,7 +207,7 @@ public class Socket: PhoenixTransportDelegate {
     self.reconnectTimer = TimeoutTimer()
     self.reconnectTimer.callback.delegate(to: self) { (self) in
       self.logItems("Socket attempting to reconnect")
-      self.teardown() { self.connect() }
+      self.teardown(reason: "reconnection") { self.connect() }
     }
     self.reconnectTimer.timerCalculation
       .delegate(to: self) { (self, tries) -> TimeInterval in
@@ -229,7 +235,12 @@ public class Socket: PhoenixTransportDelegate {
   
   /// - return: True if the socket is connected
   public var isConnected: Bool {
-    return self.connection?.readyState == .open
+    return self.connectionState == .open
+  }
+  
+  /// - return: The state of the connect. [.connecting, .open, .closing, .closed]
+  public var connectionState: PhoenixTransportReadyState {
+    return self.connection?.readyState ?? .closed
   }
   
   /// Connects the Socket. The params passed to the Socket on initialization
@@ -258,27 +269,27 @@ public class Socket: PhoenixTransportDelegate {
 //    self.connection?.enabledSSLCipherSuites = enabledSSLCipherSuites
 //    #endif
     
-    self.connection?.connect()
+    self.connection?.connect(with: self.headers)
   }
   
   /// Disconnects the socket
   ///
   /// - parameter code: Optional. Closing status code
-  /// - paramter callback: Optional. Called when disconnected
+  /// - parameter callback: Optional. Called when disconnected
   public func disconnect(code: CloseCode = CloseCode.normal,
+                         reason: String? = nil,
                          callback: (() -> Void)? = nil) {
     // The socket was closed cleanly by the User
     self.closeStatus = CloseStatus(closeCode: code.rawValue)
     
     // Reset any reconnects and teardown the socket connection
     self.reconnectTimer.reset()
-    self.teardown(code: code, callback: callback)
+    self.teardown(code: code, reason: reason, callback: callback)
   }
   
-  
-  internal func teardown(code: CloseCode = CloseCode.normal, callback: (() -> Void)? = nil) {
+  internal func teardown(code: CloseCode = CloseCode.normal, reason: String? = nil, callback: (() -> Void)? = nil) {
     self.connection?.delegate = nil
-    self.connection?.disconnect(code: code.rawValue, reason: nil)
+    self.connection?.disconnect(code: code.rawValue, reason: reason)
     self.connection = nil
     
     // The socket connection has been torndown, heartbeats are not needed
@@ -286,7 +297,7 @@ public class Socket: PhoenixTransportDelegate {
     
     // Since the connection's delegate was nil'd out, inform all state
     // callbacks that the connection has closed
-    self.stateChangeCallbacks.close.forEach({ $0.callback.call() })
+    self.stateChangeCallbacks.close.forEach({ $0.callback.call((code.rawValue, reason)) })
     callback?()
   }
   
@@ -308,10 +319,25 @@ public class Socket: PhoenixTransportDelegate {
   /// - parameter callback: Called when the Socket is opened
   @discardableResult
   public func onOpen(callback: @escaping () -> Void) -> String {
-    var delegated = Delegated<Void, Void>()
+    return self.onOpen { _ in callback() }
+  }
+  
+  /// Registers callbacks for connection open events. Does not handle retain
+  /// cycles. Use `delegateOnOpen(to:)` for automatic handling of retain cycles.
+  ///
+  /// Example:
+  ///
+  ///     socket.onOpen() { [weak self] response in
+  ///         self?.print("Socket Connection Open")
+  ///     }
+  ///
+  /// - parameter callback: Called when the Socket is opened
+  @discardableResult
+  public func onOpen(callback: @escaping (URLResponse?) -> Void) -> String {
+    var delegated = Delegated<URLResponse?, Void>()
     delegated.manuallyDelegate(with: callback)
     
-    return self.append(callback: delegated, to: &self.stateChangeCallbacks.open)
+    return self.append(callback: delegated, to: self.stateChangeCallbacks.open)
   }
   
   /// Registers callbacks for connection open events. Automatically handles
@@ -328,10 +354,27 @@ public class Socket: PhoenixTransportDelegate {
   @discardableResult
   public func delegateOnOpen<T: AnyObject>(to owner: T,
                                            callback: @escaping ((T) -> Void)) -> String {
-    var delegated = Delegated<Void, Void>()
+    return self.delegateOnOpen(to: owner) { owner, _ in callback(owner) }
+  }
+
+  /// Registers callbacks for connection open events. Automatically handles
+  /// retain cycles. Use `onOpen()` to handle yourself.
+  ///
+  /// Example:
+  ///
+  ///     socket.delegateOnOpen(to: self) { self, response in
+  ///         self.print("Socket Connection Open")
+  ///     }
+  ///
+  /// - parameter owner: Class registering the callback. Usually `self`
+  /// - parameter callback: Called when the Socket is opened
+  @discardableResult
+  public func delegateOnOpen<T: AnyObject>(to owner: T,
+                                           callback: @escaping ((T, URLResponse?) -> Void)) -> String {
+    var delegated = Delegated<URLResponse?, Void>()
     delegated.delegate(to: owner, with: callback)
     
-    return self.append(callback: delegated, to: &self.stateChangeCallbacks.open)
+    return self.append(callback: delegated, to: self.stateChangeCallbacks.open)
   }
   
   /// Registers callbacks for connection close events. Does not handle retain
@@ -346,10 +389,25 @@ public class Socket: PhoenixTransportDelegate {
   /// - parameter callback: Called when the Socket is closed
   @discardableResult
   public func onClose(callback: @escaping () -> Void) -> String {
-    var delegated = Delegated<Void, Void>()
+    return self.onClose { _, _ in callback() }
+  }
+
+  /// Registers callbacks for connection close events. Does not handle retain
+  /// cycles. Use `delegateOnClose(_:)` for automatic handling of retain cycles.
+  ///
+  /// Example:
+  ///
+  ///     socket.onClose() { [weak self] code, reason in
+  ///         self?.print("Socket Connection Close")
+  ///     }
+  ///
+  /// - parameter callback: Called when the Socket is closed
+  @discardableResult
+  public func onClose(callback: @escaping (Int, String?) -> Void) -> String {
+    var delegated = Delegated<(Int, String?), Void>()
     delegated.manuallyDelegate(with: callback)
     
-    return self.append(callback: delegated, to: &self.stateChangeCallbacks.close)
+    return self.append(callback: delegated, to: self.stateChangeCallbacks.close)
   }
   
   /// Registers callbacks for connection close events. Automatically handles
@@ -366,12 +424,28 @@ public class Socket: PhoenixTransportDelegate {
   @discardableResult
   public func delegateOnClose<T: AnyObject>(to owner: T,
                                             callback: @escaping ((T) -> Void)) -> String {
-    var delegated = Delegated<Void, Void>()
+    return self.delegateOnClose(to: owner) { owner, _ in callback(owner) }
+  }
+
+  /// Registers callbacks for connection close events. Automatically handles
+  /// retain cycles. Use `onClose()` to handle yourself.
+  ///
+  /// Example:
+  ///
+  ///     socket.delegateOnClose(self) { self, code, reason in
+  ///         self.print("Socket Connection Close")
+  ///     }
+  ///
+  /// - parameter owner: Class registering the callback. Usually `self`
+  /// - parameter callback: Called when the Socket is closed
+  @discardableResult
+  public func delegateOnClose<T: AnyObject>(to owner: T,
+                                            callback: @escaping ((T, (Int, String?)) -> Void)) -> String {
+    var delegated = Delegated<(Int, String?), Void>()
     delegated.delegate(to: owner, with: callback)
    
-    return self.append(callback: delegated, to: &self.stateChangeCallbacks.close)
+    return self.append(callback: delegated, to: self.stateChangeCallbacks.close)
   }
-  
   
   /// Registers callbacks for connection error events. Does not handle retain
   /// cycles. Use `delegateOnError(to:)` for automatic handling of retain cycles.
@@ -384,11 +458,11 @@ public class Socket: PhoenixTransportDelegate {
   ///
   /// - parameter callback: Called when the Socket errors
   @discardableResult
-  public func onError(callback: @escaping (Error) -> Void) -> String {
-    var delegated = Delegated<Error, Void>()
+  public func onError(callback: @escaping ((Error, URLResponse?)) -> Void) -> String {
+    var delegated = Delegated<(Error, URLResponse?), Void>()
     delegated.manuallyDelegate(with: callback)
     
-    return self.append(callback: delegated, to: &self.stateChangeCallbacks.error)
+    return self.append(callback: delegated, to: self.stateChangeCallbacks.error)
   }
   
   /// Registers callbacks for connection error events. Automatically handles
@@ -404,11 +478,11 @@ public class Socket: PhoenixTransportDelegate {
   /// - parameter callback: Called when the Socket errors
   @discardableResult
   public func delegateOnError<T: AnyObject>(to owner: T,
-                                            callback: @escaping ((T, Error) -> Void)) -> String {
-    var delegated = Delegated<Error, Void>()
+                                            callback: @escaping ((T, (Error, URLResponse?)) -> Void)) -> String {
+    var delegated = Delegated<(Error, URLResponse?), Void>()
     delegated.delegate(to: owner, with: callback)
 
-    return self.append(callback: delegated, to: &self.stateChangeCallbacks.error)
+    return self.append(callback: delegated, to: self.stateChangeCallbacks.error)
   }
   
   /// Registers callbacks for connection message events. Does not handle
@@ -427,7 +501,7 @@ public class Socket: PhoenixTransportDelegate {
     var delegated = Delegated<Message, Void>()
     delegated.manuallyDelegate(with: callback)
     
-    return self.append(callback: delegated, to: &self.stateChangeCallbacks.message)
+    return self.append(callback: delegated, to: self.stateChangeCallbacks.message)
   }
   
   /// Registers callbacks for connection message events. Automatically handles
@@ -447,10 +521,10 @@ public class Socket: PhoenixTransportDelegate {
     var delegated = Delegated<Message, Void>()
     delegated.delegate(to: owner, with: callback)
     
-    return self.append(callback: delegated, to: &self.stateChangeCallbacks.message)
+    return self.append(callback: delegated, to: self.stateChangeCallbacks.message)
   }
   
-  private func append<T>(callback: T, to array: inout [(ref: String, callback: T)]) -> String {
+  private func append<T>(callback: T, to array: SynchronizedArray<(ref: String, callback: T)>) -> String {
     let ref = makeRef()
     array.append((ref, callback))
     return ref
@@ -508,10 +582,10 @@ public class Socket: PhoenixTransportDelegate {
   ///
   /// - Parameter refs: List of refs returned by calls to `onOpen`, `onClose`, etc
   public func off(_ refs: [String]) {
-    self.stateChangeCallbacks.open = self.stateChangeCallbacks.open.filter({ !refs.contains($0.ref) })
-    self.stateChangeCallbacks.close = self.stateChangeCallbacks.close.filter({ !refs.contains($0.ref) })
-    self.stateChangeCallbacks.error = self.stateChangeCallbacks.error.filter({ !refs.contains($0.ref) })
-    self.stateChangeCallbacks.message = self.stateChangeCallbacks.message.filter({ !refs.contains($0.ref) })
+    self.stateChangeCallbacks.open.removeAll { refs.contains($0.ref) }
+    self.stateChangeCallbacks.close.removeAll { refs.contains($0.ref) }
+    self.stateChangeCallbacks.error.removeAll { refs.contains($0.ref) }
+    self.stateChangeCallbacks.message.removeAll { refs.contains($0.ref) }
   }
   
   
@@ -533,7 +607,8 @@ public class Socket: PhoenixTransportDelegate {
                      ref: String? = nil,
                      joinRef: String? = nil) {
     
-    let callback: (() throws -> ()) = {
+    let callback: (() throws -> ()) = { [weak self] in
+      guard let self else { return }
       let body: [Any?] = [joinRef, ref, topic, event, payload]
       let data = self.encode(body)
       
@@ -559,7 +634,7 @@ public class Socket: PhoenixTransportDelegate {
   
   /// Logs the message. Override Socket.logger for specialized logging. noops by default
   ///
-  /// - paramter items: List of items to be logged. Behaves just like debugPrint()
+  /// - parameter items: List of items to be logged. Behaves just like debugPrint()
   func logItems(_ items: Any...) {
     let msg = items.map( { return String(describing: $0) } ).joined(separator: ", ")
     self.logger?("SwiftPhoenixClient: \(msg)")
@@ -569,7 +644,7 @@ public class Socket: PhoenixTransportDelegate {
   // MARK: - Connection Events
   //----------------------------------------------------------------------
   /// Called when the underlying Websocket connects to it's host
-  internal func onConnectionOpen() {
+  internal func onConnectionOpen(response: URLResponse?) {
     self.logItems("transport", "Connected to \(endPoint)")
     
     // Reset the close status now that the socket has been connected
@@ -585,10 +660,10 @@ public class Socket: PhoenixTransportDelegate {
     self.resetHeartbeat()
     
     // Inform all onOpen callbacks that the Socket has opened
-    self.stateChangeCallbacks.open.forEach({ $0.callback.call() })
+    self.stateChangeCallbacks.open.forEach({ $0.callback.call((response)) })
   }
   
-  internal func onConnectionClosed(code: Int?) {
+  internal func onConnectionClosed(code: Int, reason: String?) {
     self.logItems("transport", "close")
 
     // Send an error to all channels
@@ -603,17 +678,17 @@ public class Socket: PhoenixTransportDelegate {
       self.reconnectTimer.scheduleTimeout()
     }
     
-    self.stateChangeCallbacks.close.forEach({ $0.callback.call() })
+    self.stateChangeCallbacks.close.forEach({ $0.callback.call((code, reason)) })
   }
   
-  internal func onConnectionError(_ error: Error) {
-    self.logItems("transport", error)
+  internal func onConnectionError(_ error: Error, response: URLResponse?) {
+    self.logItems("transport", error, response ?? "")
     
     // Send an error to all channels
     self.triggerChannelError()
     
-    // Inform any state callabcks of the error
-    self.stateChangeCallbacks.error.forEach({ $0.callback.call(error) })
+    // Inform any state callbacks of the error
+    self.stateChangeCallbacks.error.forEach({ $0.callback.call((error, response)) })
   }
   
   internal func onConnectionMessage(_ rawMessage: String) {
@@ -655,14 +730,14 @@ public class Socket: PhoenixTransportDelegate {
   
   /// Send all messages that were buffered before the socket opened
   internal func flushSendBuffer() {
-    guard isConnected && sendBuffer.count > 0 else { return }
+    guard isConnected else { return }
     self.sendBuffer.forEach( { try? $0.callback() } )
-    self.sendBuffer = []
+    self.sendBuffer.removeAll()
   }
   
   /// Removes an item from the sendBuffer with the matching ref
   internal func removeFromSendBuffer(ref: String) {
-    self.sendBuffer = self.sendBuffer.filter({ $0.ref != ref })
+    self.sendBuffer.removeAll { $0.ref == ref }
   }
 
   /// Builds a fully qualified socket `URL` from `endPoint` and `params`.
@@ -721,13 +796,13 @@ public class Socket: PhoenixTransportDelegate {
     // Do not start up the heartbeat timer if skipHeartbeat is true
     guard !skipHeartbeat else { return }
 
-    self.heartbeatTimer = HeartbeatTimer(timeInterval: heartbeatInterval)
+    self.heartbeatTimer = HeartbeatTimer(timeInterval: heartbeatInterval, leeway: heartbeatLeeway)
     self.heartbeatTimer?.start(eventHandler: { [weak self] in
       self?.sendHeartbeat()
     })
   }
   
-  /// Sends a hearbeat payload to the phoenix serverss
+  /// Sends a heartbeat payload to the phoenix servers
   @objc func sendHeartbeat() {
     // Do not send if the connection is closed
     guard isConnected else { return }
@@ -774,21 +849,21 @@ public class Socket: PhoenixTransportDelegate {
   //----------------------------------------------------------------------
   // MARK: - TransportDelegate
   //----------------------------------------------------------------------
-  public func onOpen() {
-    self.onConnectionOpen()
+  public func onOpen(response: URLResponse?) {
+    self.onConnectionOpen(response: response)
   }
   
-  public func onError(error: Error) {
-    self.onConnectionError(error)
+  public func onError(error: Error, response: URLResponse?) {
+    self.onConnectionError(error, response: response)
   }
   
   public func onMessage(message: String) {
     self.onConnectionMessage(message)
   }
   
-  public func onClose(code: Int) {
+  public func onClose(code: Int, reason: String? = nil) {
     self.closeStatus.update(transportCloseCode: code)
-    self.onConnectionClosed(code: code)
+    self.onConnectionClosed(code: code, reason: reason)
   }
 }